requirements.txt
openai~=1.75.0
python-dotenv~=1.1.0
PyPDF2~=3.0.1
numpy~=2.2.5
faiss-cpu~=1.10.0